ms-entrenador y clase se agrega verificacion en service delete

+++ ms-clase/src/main/java/com/example/ms_clase/service/ClaseService.java
@@ -80,6 +80,9 @@ public class ClaseService {
     }
 
     public void delete(Long id ){
+        if (!repository.existsById(id)) {
+            throw new EntityNotFoundException("No se puede eliminar clase no encontrada");
+        }
         repository.deleteById(id);
 
     }
